+++ web-app/src/ng1/factories/map.service.js
@@ -173,6 +173,7 @@ function MapService(EventService, LocationService, FeatureService, $compile, $ro
       } else if (layer.type === 'Feature') {
         FeatureService.getFeatureCollection(event, layer).then(function(featureCollection) {
           service.createVectorLayer({
+            id: layer.id,
             name: layer.name, // TODO need to track by id as well not just names
             group: 'feature',
             type: 'geojson',