+++ CHANGELOG.md
@@ -9,6 +9,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 * Added support in configuration/environment for MongoDB replica sets.
 
 ##### Bug Fixes
+* Docer build now works with Angular CLI
 
 ## [5.3.5](https://github.com/ngageoint/mage-server/releases/tag/5.3.5)