+++ CHANGELOG.md
@@ -14,6 +14,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 * Hide admin icon in navbar for non admin users.
 * New users created by admin should default to 'active'.
 * Docker build now works with Angular CLI
+* Fixed swagger page.
 
 ## [5.3.5](https://github.com/ngageoint/mage-server/releases/tag/5.3.5)