+++ CHANGELOG.md
@@ -6,6 +6,13 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 ## Pending on [`develop`](https://github.com/ngageoint/mage-server/tree/develop)
 
 ##### Features
+
+##### Bug Fixes
+
+## [5.4.0](https://github.com/ngageoint/mage-server/releases/tag/5.4.0)
+
+##### Features
+* Added pagination for users, teams, events, devices and layers. This will greatly decrease load times for admin pages on servers with more users.
 * Added observation location accuracy.
 * Added support in configuration/environment for MongoDB replica sets.